// mobile/screens/health/admin/admin-childhealth/forms/muti-step-form/types.ts
export interface ChildHealthRecord {
  chrec_id: number;
  pat_id: string;
  fname: string;
  lname: string;
  mname: string;
  sex: string;
  age: string;
  dob: string;
  householdno: string;
  street: string;
  sitio: string;
  barangay: string;
  city: string;
  province: string;
  landmarks: string;
  pat_type: string;
  address: string;
  mother_fname: string;
  mother_lname: string;
  mother_mname: string;
  mother_contact: string;
  mother_occupation: string;
  father_fname: string;
  father_lname: string;
  father_mname: string;
  father_contact: string;
  father_occupation: string;
  family_no: string;
  birth_weight: number;
  birth_height: number;
  type_of_feeding: string;
  delivery_type: string;
  place_of_delivery_type: string;
  pod_location: string;
  pod_location_details?: string;
  health_checkup_count: number;
  birth_order?: string;
  tt_status?: string; // Optional field for TT status
}